+++ packages/node/src/remote/client.ts
@@ -12,6 +12,7 @@ import { FetchOptions } from '../types/fetch';
 import { ExperimentUser } from '../types/user';
 import { Variant, Variants } from '../types/variant';
 import { populateRemoteConfigDefaults } from '../util/config';
+import { AmplitudeLogger } from '../util/logger';
 import { sleep } from '../util/time';
 import {
   evaluationVariantsToVariants,
@@ -23,6 +24,7 @@ import {
  * @category Core Usage
  */
 export class RemoteEvaluationClient {
+  private readonly logger: AmplitudeLogger;
   private readonly apiKey: string;
   private readonly config: RemoteEvaluationConfig;
   private readonly evaluationApi: EvaluationApi;
@@ -36,6 +38,10 @@ export class RemoteEvaluationClient {
   public constructor(apiKey: string, config?: RemoteEvaluationConfig) {
     this.apiKey = apiKey;
     this.config = populateRemoteConfigDefaults(config);
+    this.logger = new AmplitudeLogger(
+      this.config.logLevel,
+      this.config.loggerProvider,
+    );
     this.evaluationApi = new SdkEvaluationApi(
       apiKey,
       this.config.serverUrl,
@@ -62,7 +68,7 @@ export class RemoteEvaluationClient {
     if (!this.apiKey) {
       throw Error('Experiment API key is empty');
     }
-    this.debug('[Experiment] Fetching variants for user: ', user);
+    this.logger.debug('[Experiment] Fetching variants for user: ', user);
     try {
       return await this.doFetch(user, this.config.fetchTimeoutMillis, options);
     } catch (e) {
@@ -125,7 +131,7 @@ export class RemoteEvaluationClient {
       userContext,
       getVariantsOptions,
     );
-    this.debug('[Experiment] Fetched variants: ', results);
+    this.logger.debug('[Experiment] Fetched variants: ', results);
     return evaluationVariantsToVariants(results);
   }
 
@@ -134,7 +140,7 @@ export class RemoteEvaluationClient {
     options: FetchOptions,
     err: Error,
   ): Promise<Record<string, Variant>> {
-    this.debug('[Experiment] Retrying fetch');
+    this.logger.debug('[Experiment] Retrying fetch');
     let delayMillis = this.config.fetchRetryBackoffMinMillis;
     for (let i = 0; i < this.config.fetchRetries; i++) {
       await sleep(delayMillis);
@@ -161,13 +167,6 @@ export class RemoteEvaluationClient {
       library: `experiment-node-server/${PACKAGE_VERSION}`,
       ...user,
     };
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  private debug(message?: any, ...optionalParams: any[]): void {
-    if (this.config.debug) {
-      console.debug(message, ...optionalParams);
-    }
   }
 
   private shouldRetryFetch(e: Error): boolean {